+++ app/components/ui/index.ts
@@ -1,4 +1,4 @@
 export * from './BrandSymbol';
 export * from './Button';
 export * from './Card';
-export * from './Link';
+export * from './LinkButton';
